+++ packages/json/CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
-## 0.10.
+## 0.10.0 — not yet released
 
-* Add `LionWebJsonDeltaChunk` type, specifically for the delta protocol.
+* Add a `LionWebJsonDeltaChunk` type for the delta protocol.
+    (Also add a `OnlyNodesOfLionWebJsonChunk` type to refactor to DRY.)
 
 
 ## 0.9.0